Crops cells out of image, centre, pad with 0s
Given the bounding box and binary mask for a cell,
crop it out of the image it's in, optionally having
the cell's background padded with zeros, so that
other cells are removed.

Another function takes this crop and centres it in
a square image, resulting in a small image, in which
a cell sits near the middle, with a black backgroun.